# openapi.yaml
openapi: 3.0.0
info:
  title: WhatsApp Automation Bot API
  description: Production-ready WhatsApp automation bot system with API-like features powered by Playwright
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://your-domain.com/api
    description: Production server
security:
  - ApiKeyAuth: []
  - BearerAuth: []
...
paths:
...
  /webhook:
    get:
      tags:
        - Webhooks
      summary: Get webhooks
      description: Get all registered webhooks
      responses:
        '200':
          description: Webhooks retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Webhook'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /webhook/register:
    post:
      tags:
        - Webhooks
      summary: Register webhook
      description: Register a new webhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - url
                - events
              properties:
                url:
                  type: string
                  format: uri
                  description: Webhook URL
                  example: "https://your-webhook-url.com/whatsapp"
                events:
                  type: array
                  description: Events to subscribe to
                  items:
                    type: string
                    enum: [message.new, message.sent, session.connected, session.disconnected]
                  example: ["message.new", "message.sent"]
                secret:
                  type: string
                  description: Webhook secret for verification
                  example: "your-webhook-secret"
      responses:
        '201':
          description: Webhook registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Webhook registered successfully"
                  webhookId:
                    type: integer
                    example: 1
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /webhook/unregister/{id}:
    delete:
      tags:
        - Webhooks
      summary: Unregister webhook
      description: Unregister a webhook
      parameters:
        - name: id
          in: path
          required: true
          description: Webhook ID
          schema:
            type: integer
      responses:
        '200':
          description: Webhook unregistered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Webhook unregistered successfully"
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /webhook/test:
    post:
      tags:
        - Webhooks
      summary: Test webhook
      description: Test a webhook endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - url
                - event
              properties:
                url:
                  type: string
                  format: uri
                  description: Webhook URL to test
                  example: "https://your-webhook-url.com/whatsapp"
                event:
                  type: string
                  description: Event to test
                  example: "message.new"
                secret:
                  type: string
                  description: Webhook secret
                  example: "your-webhook-secret"
      responses:
        '200':
          description: Webhook test completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Webhook test sent successfully"
                  response:
                    type: object
                    properties:
                      status:
                        type: integer
                        example: 200
                      statusText:
                        type: string
                        example: "OK"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
...
  schemas:
...
    Webhook:
      type: object
      properties:
        id:
          type: integer
          example: 1
        url:
          type: string
          format: uri
          example: "https://your-webhook-url.com/whatsapp"
        events:
          type: array
          items:
            type: string
          example: ["message.new", "message.sent"]
        is_active:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
          example: "2024-01-01T12:00:00.000Z"
        updated_at:
          type: string
          format: date-time
          example: "2024-01-01T12:00:00.000Z"
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: "Validation failed"
              details:
                type: array
                items:
                  type: object
                  properties:
                    field:
                      type: string
                    message:
                      type: string
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: "Invalid or expired token"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: "Resource not found"
...